cybersec-web/01_ft_otp/requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
crypto==1.4.1
cryptography==46.0.3
idna==3.11
Naked==0.1.32
pycparser==2.23
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
shellescape==3.8.1
urllib3==2.5.0
